Create serializers for db models
Create serializer for product, tenant, category, customer and provider.

Only the main serializer and the create serializer for each one of this where created. A sereializer for invoice and roles were not created.

Roles will be handled in another PR.

This serializers are a starting point and may change in the future.